+++ History.md
@@ -1,3 +1,17 @@
+0.5.0 / 2014-04-22 
+==================
+
+ * `shouldLoadMediaGen` option, used by pjs-player. If false, the `READY` event will fire when the config is loaded.
+ * Never use default error message for media gen error responses.
+ * Use default error message unless `verboseErrorMessaging:true` is passed in constructor.
+ * Use overrideParams in mediaGen request. 
+ * Add test for timedTextURL.
+ * Use live PMT url. 
+ * Handle mediaGen errors.
+ * Append overaly item to mediaGen.
+ * Update tests now that config response returns correctly formatted errors.
+ * Set request header `Accept` to application/json. When receiving a json error response, dispatch the `ERROR` event with the message in the response.
+ * Update tests to use media-resolver.
 
 0.4.0 / 2014-04-01 
 ==================